feat(sidebars-mutations): rename
